Type HMR
